requirements.txt
autogen_ext
autogen_core
tiktoken
anthropic
vcdvcd
pyvcd
python-dotenv